NSM-143  missing file message during delay dl, email if error during dl, discrete site daily flow option, logging, moved one close to finally from catch

+++ src/main/java/gov/cida/sedmap/io/InputStreamWithFile.java
@@ -28,6 +28,8 @@ public class InputStreamWithFile extends InputStream {
 	public void deleteFile() {
 		IoUtils.quiteClose(this);
 
+		if (file == null) return;
+
 		if ( ! file.delete() ) {
 			file.deleteOnExit();
 		}
